change row desgin for task adapter
