+++ app/src/main/AndroidManifest.xml
@@ -44,13 +44,13 @@
 
         <!-- addingTeacher activity -->
         <activity
-            android:name=".AddingTeacher"
+            android:name=".activity.AddingTeacher"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
 
         <!-- students activity -->
         <activity
-            android:name=".StudentsActivity"
+            android:name=".activity.StudentsActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
 
@@ -62,7 +62,7 @@
 
         <!-- missions activity -->
         <activity
-            android:name=".Missions"
+            android:name=".activity.TaskActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
