+++ app/src/main/java/com/example/quranapp/activity/StudentsActivity.java
@@ -1,4 +1,4 @@
-package com.example.quranapp;
+package com.example.quranapp.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.SearchView;
@@ -8,6 +8,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import android.app.SearchManager;
 import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
@@ -18,6 +19,13 @@ import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.quranapp.AddingStudent;
+import com.example.quranapp.R;
+import com.example.quranapp.model.Student;
+import com.example.quranapp.adapter.StudentAdapter;
+import com.example.quranapp.URLs;
+import com.example.quranapp.ViewDialog;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
 
 import org.json.JSONArray;
@@ -27,65 +35,82 @@ import org.json.JSONObject;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Missions extends AppCompatActivity implements MissionsAdapter.MissionsAdapterListener {
-    private MissionsAdapter adapter;
+public class StudentsActivity extends AppCompatActivity implements StudentAdapter.StudentAdapterListener {
+
+    private StudentAdapter adapter;
     private JSONArray jsonArray;
-    List<Task> listItems ;
+
+    List<Student> listItems ;
+    ViewDialog viewDialog;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.missions_activity);
-        Toolbar toolbar = findViewById(R.id.missions_toolbar);
+        setContentView(R.layout.students_activity);
 
+        Toolbar toolbar = findViewById(R.id.students_toolbar);
         setSupportActionBar(toolbar);
-        RecyclerView recyclerView = findViewById(R.id.missions_recyclerView);
+        viewDialog = new ViewDialog(this);
 
+        RecyclerView recyclerView = findViewById(R.id.students_recyclerView);
         recyclerView.setHasFixedSize(true);
 
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         listItems = new ArrayList<>();
 
-        adapter = new MissionsAdapter(listItems, this);
+        FloatingActionButton studentsFAB = findViewById(R.id.students_fab);
+        studentsFAB.setOnClickListener(v -> {
+            Intent i = new Intent(getApplicationContext(), AddingStudent.class);
+            startActivity(i);
+        });
+
+
+        adapter = new StudentAdapter(listItems, this);
         recyclerView.setAdapter(adapter);
 
-        getTasks();
-        // toolbar fancy stuff
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setTitle(R.string.toolbar_title);
+
+        toolbar.setNavigationOnClickListener(v -> {
+            finish();
+        });
+
+
+        getStudents();
     }
 
-    private void getTasks() {
-        RequestQueue requestQueue = Volley.newRequestQueue(this);
-        StringRequest stringRequest = new StringRequest(Request.Method.GET, URLs.GetSTasks  , response -> {
+
+    public void getStudents(){
+        viewDialog.showDialog();
+        RequestQueue requestQueue = Volley.newRequestQueue(getApplicationContext());
+        StringRequest stringRequest = new StringRequest(Request.Method.GET, URLs.GetStudents, response -> {
             try {
                 jsonArray = new JSONArray(response);
                 for (int i = 0; i < jsonArray.length(); i ++){
-                    JSONObject TaskObject = jsonArray.getJSONObject(i);
-                    int Id = TaskObject.getInt("IdTask");
-                    String TaskName = TaskObject.getString("TaskName");
-                    String TaskDec = TaskObject.getString("TaskDec");
-                    String CreatedAt = TaskObject.getString("CreatedAt");
-                    Task listItem = new Task(Id,TaskName, TaskDec, CreatedAt);
+                    JSONObject StudentObject = jsonArray.getJSONObject(i);
+                    String Name = StudentObject.getString("Name");
+                    String PhoneNumber = StudentObject.getString("PhoneNumber");
+                    String Date = StudentObject.getString("CreatedAt");
+                    Student listItem = new Student(Name, PhoneNumber, Date);
                     listItems.add(listItem);
                 }
 
                 adapter.notifyDataSetChanged();
-//                viewDialog.hideDialog();
+                viewDialog.hideDialog();
                 Log.d("res", jsonArray.toString());
 
             } catch (JSONException e) {
                 e.printStackTrace();
-//                viewDialog.hideDialog();
-                Snackbar.make(findViewById(android.R.id.content), "Couldn't get Students " + e , Snackbar.LENGTH_LONG)
-                        .setAction("Retry", v -> getTasks()).show();
+                viewDialog.hideDialog();
+                Snackbar.make(findViewById(android.R.id.content), " تعذر عرض الدارسين " + e , Snackbar.LENGTH_LONG)
+                        .setAction("محاولة مرة اخري", v -> getStudents()).show();
             }
 
         }, error -> {
             error.printStackTrace();
-//            viewDialog.hideDialog();
-            Snackbar.make(findViewById(android.R.id.content),"Couldn't get Students " + error , Snackbar.LENGTH_LONG)
-                    .setAction("Retry", v -> getTasks()).show();
+            viewDialog.hideDialog();
+            Snackbar.make(findViewById(android.R.id.content), " تعذر عرض الدارسين " + error , Snackbar.LENGTH_LONG)
+                    .setAction("محاولة مرة اخري", v -> getStudents()).show();
         });
 
         requestQueue.add(stringRequest);
@@ -114,6 +139,7 @@ public class Missions extends AppCompatActivity implements MissionsAdapter.Missi
             @Override
             public boolean onQueryTextChange(String newText) {
                 adapter.getFilter().filter(newText);
+
                 return false;
             }
         });
@@ -122,8 +148,8 @@ public class Missions extends AppCompatActivity implements MissionsAdapter.Missi
     }
 
     @Override
-    public void onContactSelected(Teacher teacher) {
-        Toast.makeText(getApplicationContext(), "Selected: " + teacher.getName() + ", " + teacher.getPhone(), Toast.LENGTH_LONG).show();
+    public void onStudentSelected(Student student) {
+        Toast.makeText(getApplicationContext(), "Selected: " + student.getName() + ", " + student.getPhone(), Toast.LENGTH_LONG).show();
     }
 }
