+++ app/src/main/java/com/example/quranapp/adapter/StudentAdapter.java
@@ -1,4 +1,4 @@
-package com.example.quranapp;
+package com.example.quranapp.adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -11,33 +11,35 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.quranapp.R;
+import com.example.quranapp.model.Student;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class MissionsAdapter extends RecyclerView.Adapter<MissionsAdapter.ViewHolder> implements Filterable {
-    private List<Task> listItems;
-    private List<Task> listItemsFiltered;
+public class StudentAdapter extends RecyclerView.Adapter<StudentAdapter.ViewHolder> implements Filterable {
+    private List<Student> listItems;
+    private List<Student> listItemsFiltered;
 
-
-    public MissionsAdapter(List<Task> listItems, Context context) {
+    public StudentAdapter(List<Student> listItems, Context context) {
         this.listItems = listItems;
         listItemsFiltered = new ArrayList<>(listItems);
-
     }
 
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.row_mission, parent, false);
+        View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.row_student, parent, false);
         return new ViewHolder(v);
     }
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        Task listItem = listItems.get(position);
-        holder.textViewName.setText(listItem.getTaskName());
-        holder.textViewPhone.setText(listItem.getTaskDec());
-        holder.textViewDate.setText(listItem.getCreatedAt());
+        Student listItem = listItems.get(position);
+        holder.textStudentName.setText(listItem.getName());
+        holder.textStudentAddress.setText(listItem.getName());
+        holder.textStudentPhoneNumber.setText(listItem.getPhone());
+        holder.timestamp.setText(listItem.getDate());
     }
 
     @Override
@@ -45,35 +47,39 @@ public class MissionsAdapter extends RecyclerView.Adapter<MissionsAdapter.ViewHo
         return listItems.size();
     }
 
+    public class ViewHolder extends RecyclerView.ViewHolder {
+        public TextView textStudentName;
+        public TextView textStudentAddress;
+        public TextView textStudentPhoneNumber;
+        public TextView timestamp;
+
+        public ViewHolder(@NonNull View itemView) {
+            super(itemView);
+
+            textStudentName = itemView.findViewById(R.id.textStudentName);
+            textStudentAddress = itemView.findViewById(R.id.textteacherName);
+            textStudentPhoneNumber = itemView.findViewById(R.id.textTaskDetails);
+            timestamp = itemView.findViewById(R.id.timestamp);
+
+        }
+    }
+
     @Override
     public Filter getFilter() {
         return contactFilter;
     }
 
-    public class ViewHolder extends RecyclerView.ViewHolder {
-        public TextView textViewName;
-        public TextView textViewPhone;
-        public TextView textViewDate;
-
-        public ViewHolder(@NonNull View itemView) {
-            super(itemView);
-
-            textViewName = itemView.findViewById(R.id.mission_row_name_textView);
-            textViewPhone = itemView.findViewById(R.id.mission_row_phone_textView);
-            textViewDate = itemView.findViewById(R.id._mission_row_date_textView);
-        }
-    }
     private Filter contactFilter = new Filter() {
         @Override
         protected FilterResults performFiltering(CharSequence constraint) {
-            List<Task> filteredList =new ArrayList<>();
+            List<Student> filteredList =new ArrayList<>();
 
             if (constraint==null|| constraint.length()==0){
                 filteredList.addAll(listItemsFiltered);
             }else {
                 String filterPattern = constraint.toString().toLowerCase().trim();
-                for (Task item : listItemsFiltered){
-                    if (item.getTaskName().toLowerCase().contains(filterPattern) || item.getTaskDec().contains(filterPattern)){
+                for (Student item : listItemsFiltered){
+                    if (item.getName().toLowerCase().contains(filterPattern) || item.getPhone().contains(filterPattern)){
                         filteredList.add(item);
                     }
                 }
@@ -91,9 +97,7 @@ public class MissionsAdapter extends RecyclerView.Adapter<MissionsAdapter.ViewHo
 
         }
     };
-
-    public interface MissionsAdapterListener {
-        void onContactSelected(Teacher teacher);
+    public interface StudentAdapterListener {
+        void onStudentSelected(Student student);
     }
-
 }
