--- app/src/main/res/layout/row_mission.xml
@@ -1,108 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:background="?attr/selectableItemBackground"
-    android:clickable="true"
-    android:focusable="true">
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/mission_cardView"
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:layout_marginVertical="4dp"
-        android:background="#ffffff"
-        app:cardCornerRadius="15dp"
-        android:clickable="true"
-        android:focusable="true">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="97dp"
-            android:layout_gravity="center_horizontal|center_vertical"
-            android:layout_margin="4dp"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:id="@+id/mission_row_imageView"
-                android:layout_width="100dp"
-                android:layout_height="95dp"
-                android:layout_gravity="center_horizontal"
-                android:layout_marginEnd="4dp"
-                android:layout_marginRight="4dp"
-                android:src="@drawable/img_card"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:id="@+id/mission_row_name_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginVertical="20dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginEnd="8dp"
-                android:layout_marginRight="8dp"
-                android:gravity="center"
-                android:textAlignment="center"
-                android:textColor="#557571"
-                android:textSize="12sp"
-                android:textStyle="bold"
-                app:layout_constraintEnd_toStartOf="@+id/mission_row_imageView"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:id="@+id/mission_row_phone_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginVertical="20dp"
-                android:layout_marginTop="8dp"
-                android:layout_marginEnd="8dp"
-                android:layout_marginRight="8dp"
-                android:gravity="center"
-                android:textAlignment="center"
-                android:textColor="#557571"
-                android:textSize="12sp"
-                app:layout_constraintEnd_toStartOf="@+id/mission_row_imageView"
-                app:layout_constraintTop_toBottomOf="@+id/mission_row_name_textView" />
-
-            <TextView
-                android:id="@+id/_mission_row_date_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginVertical="20dp"
-                android:layout_marginStart="4dp"
-                android:layout_marginLeft="4dp"
-                android:gravity="center"
-                android:textAlignment="center"
-                android:textColor="#557571"
-                android:textSize="12sp"
-                android:textStyle="bold"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/mission_row_status_textView" />
-
-            <TextView
-                android:id="@+id/mission_row_status_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:layout_marginVertical="20dp"
-                android:layout_marginStart="4dp"
-                android:layout_marginLeft="4dp"
-                android:text="غير مكتملة"
-                android:gravity="center"
-                android:textAlignment="center"
-                android:textColor="#F30909"
-                android:textSize="12sp"
-                android:textStyle="italic"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </androidx.cardview.widget.CardView>
-</RelativeLayout>
